% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildup.R
\name{build_up}
\alias{build_up}
\title{build_up() build up method for calculating Air Exchange Rate. This function is the objective function whose
roots we want to find with respect to the Air Exchange Rate}
\usage{
build_up(aer, n, Gp, V, C0, C1, delta_t, Cr = 400)
}
\arguments{
\item{aer}{: Air Exchange Rate ( 1 / hour)}

\item{n}{: number of individuals}

\item{Gp}{: Average age-adjusted CO2 generation rate (L / (min·person))}

\item{V}{: Room volume in meters cubed}

\item{C0}{: Initial CO2 concentration (ppm)}

\item{C1}{: Final CO2 concentration (ppm)}

\item{delta_t}{: Time difference (hours)}

\item{Cr}{: Replacement air CO2 concentration. Will set to a default value of ambient air (400 ppm)}
}
\value{
: Return the value of the equation given the input parameters
}
\description{
build_up() build up method for calculating Air Exchange Rate. This function is the objective function whose
roots we want to find with respect to the Air Exchange Rate
}
