% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_up_functions.R
\name{build_up_main}
\alias{build_up_main}
\title{build_up_main() Highest level function call to estimate parameters using the secant method for the
build up equation.}
\usage{
build_up_main(
  time_vec,
  co2_vec,
  aer = NA,
  n = NA,
  Gp = NA,
  V = NA,
  n_change = NA,
  quant_cutoff = 0.9,
  Cr = 400,
  tol = 1e-10,
  max_iter = 1000
)
}
\arguments{
\item{time_vec}{: Vector of time values. Must have POSIXct attribute. Must call in vectors
from dataframes using the $ operator so that the data keeps the POSIXct attribute}

\item{co2_vec}{: Vector of CO2 values. Must be in ppm, although there is no way for the
program to check this}

\item{aer}{: Air Exchange Rate ( 1 / hour). Can be input as a single value or vector, depending on size of n.}

\item{n}{: number of individuals. If known, will either be input as a number. This will mean that n is
assumed to be constant. If time varying, then n will be a mxp matrix. Each time that n changes, another
row will be added. Thus, with x changes in n, m = x + 1. p will represent each type of individual in the
room. n[m,p] will give the number of individuals of type p for the current population.}

\item{Gp}{: Average age-adjusted CO2 generation rate (L / (min·person)). The length of this vector must be equal
to either the number of columns of n, rows of n, or be a matrix of the same size of n, depending on the input of n.}

\item{V}{: Room volume in meters cubed. Can be input as a single value or vector, depending on size of n.}

\item{n_change}{: This vector is a vector of time values. Must have POSIXct attribute. Must call in vectors
from dataframes using the $ operator so that the data keeps the POSIXct attribute. This vector will represent
the times at which the number of people in the room change. If n = NA, will make it so that n_change must be
NA as well. If n is not constant, the length of n_change must be equal to the number of rows of n. This means that
the first input of n_change should always be the start of the time series values.
#'@param quant_cutoff : Percentile cutoff for identifying exponential periods. Variance within
the CO2 values p index values to the left and right is calculated. High values of variance indicate
an exponential period. Percentile based cutoff is made to identify these high values, and therefore,
the exponential periods. A higher value for this cutoff is more strict, and may identify less exponential
periods. A lower values for this cutoff is less strict, and may identify more exponential periods.}

\item{Cr}{: Replacement Air CO2 concentration (ppm). Will set to a default value of ambient air (400 ppm)}

\item{tol}{: absolute difference of function values between steps we use as a stopping condition}

\item{max_iter}{: maximum number of iterations}
}
\value{
: A list with the following attributes:
       *summary_raw : returns the summary() of the esimated parameter using raw CO2 data
       *summary_interval : returns the summary() of the esimated parameter using interval average CO2 data
       *raw_co2_df : returns the period_data_final data frame. Useful to get the distribution of the parameter
       estimates, and additional information on regions of estimation and diagnostic information from
       the root finding method.
       *interval_avg_df : returns the period_data_final_ia data frame. May be useful when ranges of CO2 are 
       small and error from sensor readings can impair analysis with raw data
       *parameters : returns information on the parameter arguments used while calling build_up_method
       Each row corresponds to the chronological change in parameters. Will always be at least one row,
       which represents the parameters at time = 0.
       removed_periods : Base return value is a string stating that no periods were removed. For periods
       in which 0 values will be passed as arguments, these periods are removed from period_data and
       stored here for troubleshooting. 0 arguments are nonsensical and can not be handled by this method.
       interval_p_value : Returns the value of p used for calculating the interval CO2 values
}
\description{
build_up_main() Highest level function call to estimate parameters using the secant method for the
build up equation.
}
