% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_up_functions.R
\name{build_up_method}
\alias{build_up_method}
\title{build_up_method() build up method for estimating parameters. Estimates parameter value by solving for the
root of the build_up() function}
\usage{
build_up_method(
  aer = NA,
  n = NA,
  Gp = NA,
  V = NA,
  C0,
  C1,
  delta_t,
  Cr = 400,
  tol = 1e-10,
  max_iter = 1000
)
}
\arguments{
\item{aer}{: Air Exchange Rate ( 1 / hour)}

\item{n}{: number of individuals}

\item{Gp}{: Average age-adjusted CO2 generation rate (L / (min·person))}

\item{V}{: Room volume in meters cubed}

\item{C0}{: Initial CO2 concentration (ppm)}

\item{C1}{: Final CO2 concentration (ppm)}

\item{delta_t}{: Time difference (hours)}

\item{Cr}{: Replacement Air CO2 concentration (ppm). Will set to a default value of ambient air (400 ppm)}

\item{tol}{: absolute difference of function values between steps we use as a stopping condition}

\item{max_iter}{: maximum number of iterations}
}
\value{
: A list containing the following attributes:
   * root - Parameter value with build_up(parameter) close to zero
   * f_root - build_up(root)
   * iter - number of iterations to reach the solution
   * convergence - 0 if the root was found successfully, 1 if not found
}
\description{
build_up_method() build up method for estimating parameters. Estimates parameter value by solving for the
root of the build_up() function
}
