% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO2rate_to_Gp.R
\name{CO2rate_to_Gp}
\alias{CO2rate_to_Gp}
\title{Convert CO2 emission rates (in micrograms/s) to Gp (in L/min)}
\usage{
CO2rate_to_Gp(CO2rate, temp = 25)
}
\arguments{
\item{CO2rate}{}

\item{temp}{}
}
\value{
Gp
}
\description{
Convert CO2 emission rates (in micrograms/s) to Gp (in L/min)
}
