% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/estimate_ventilation.R
\name{estimate_ventilation}
\alias{estimate_ventilation}
\title{Estimate Ventilation Rate}
\usage{
estimate_ventilation(
  freq,
  CO2,
  volume,
  envCO2known = NULL,
  init.Q,
  temp = 25,
  persondata = NULL,
  critpoints = NULL,
  method = "NR",
  max.iter = 1000,
  tol = 1e-08,
  E.init = NULL,
  envCO2.init = NULL,
  verbose = FALSE,
  record.steps = FALSE
)
}
\arguments{
\item{freq}{The frequency of CO2 measurements (e.g., measurements per hour).}

\item{CO2}{A numeric vector representing the CO2 concentrations over time.}

\item{volume}{The volume of the room in cubic meters.}

\item{envCO2known}{Known environmental CO2 concentration in ug/m^3.}

\item{init.Q}{Initial estimate of ventilation rate.}

\item{temp}{The temperature in degrees Celsius (default is 25).}

\item{persondata}{A data frame containing information about individuals in the room, used for Newton-Raphson optimization.}

\item{critpoints}{Critical points for multi-dimensional Newton method.}

\item{method}{The optimization method to use ('NR' for Newton-Raphson, 'Newton' for multi-dimensional Newton).}

\item{max.iter}{Maximum number of iterations for optimization.}

\item{tol}{Tolerance for convergence in optimization.}

\item{E.init}{Initial estimate of emissions rate for multi-dimensional Newton method.}

\item{envCO2.init}{Initial estimate of environmental CO2 for multi-dimensional Newton method.}

\item{verbose}{If TRUE, print iteration details during optimization.}

\item{record.steps}{If TRUE, record optimization steps.}
}
\value{
A list containing the estimated ventilation rate, number of iterations, and convergence status.
}
\description{
This function estimates ventilation rates in indoor environments based on CO2 time-series data.
}
\examples{
# Example using Newton-Raphson method
estimate_ventilation(freq = 1, CO2 = c(400, 600, 800), volume = 100,
                     envCO2known = 500, init.Q = 1, temp = 25,
                     persondata = data.frame(time = c(0, 1, 2), activity = c(1, 2, 1)),
                     method = 'NR')

# Example using multi-dimensional Newton method
estimate_ventilation(freq = 1, CO2 = c(400, 600, 800), volume = 100,
                     envCO2known = 500, init.Q = 1, temp = 25,
                     persondata = data.frame(time = c(0, 1, 2), activity = c(1, 2, 1)),
                     method = 'Newton')

}
