% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example_data.R
\docType{data}
\name{example_data}
\alias{example_data}
\title{Example data set.}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1652 rows and 5 columns.
}
\usage{
example_data
}
\description{
Example data set.
}
\keyword{datasets}
