\name{Gp_rate}
\alias{Gp_rate}
\title{Estimate Gp from Respiratory Rate and CO2 Concentration in Breath}
\description{
  Calculates the CO2 generation rate (Gp) based on the concentration of CO2 in exhaled breath and the respiratory rate of an individual. This function is particularly useful in environmental studies where respiratory factors influence CO2 levels.
}
\usage{
Gp_rate(resp_rate, co2_breath = 4e4)
}
\arguments{
  \item{resp_rate}{Respiratory rate of the individual in cubic meters per hour.}
  \item{co2_breath}{Concentration of CO2 in exhaled breath in parts per million (ppm). Default value is set to 40000 ppm.}
}
\details{
  The function utilizes respiratory rate and the CO2 concentration in exhaled breath to estimate Gp. The default CO2 concentration in breath is based on standard values but can be adjusted as needed.
}
\value{
  Returns the calculated Gp value, indicating the CO2 generation rate based on respiratory factors.
}
\references{
  %% Add references to literature or websites related to respiratory physiology and environmental studies here %%
}
\author{
  %% Your Name / Your Institution %%
}
\note{
  For accurate Gp estimation, it is recommended to use respiratory rate values from reliable sources like the EPA Exposure Factors Handbook.
}
\seealso{
  \code{\link{Gp_met}}, \code{\link{another_complementary_function}}
}
\examples{
## Example of using Gp_rate
## Not run:
# Gp_rate(resp_rate = 0.5, co2_breath = 4e4)
## End(Not run)
}
\keyword{respiratory}
\keyword{environment}
