% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO2rate_to_Gp.R
\name{Gp_to_CO2rate}
\alias{Gp_to_CO2rate}
\title{Convert Gp (in L/min) to CO2 emission rates (in milligrams/s)}
\usage{
Gp_to_CO2rate(Gp, temp = 25)
}
\arguments{
\item{temp}{}
}
\value{
CO2rate
}
\description{
Convert Gp (in L/min) to CO2 emission rates (in milligrams/s)
}
