% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO2rate_to_Gp.R
\name{mg_to_ppm}
\alias{mg_to_ppm}
\title{convert from mg/m^3 to ppm}
\usage{
mg_to_ppm(mg, temp)
}
\arguments{
\item{mg}{concentration of CO2 in mg/m^3}

\item{temp}{temperature in celsius}
}
\description{
convert from mg/m^3 to ppm
}
