% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CO2rate_to_Gp.R
\name{ppm_to_mg}
\alias{ppm_to_mg}
\title{Convert from ppm to mg/m^3}
\usage{
ppm_to_mg(ppm, temp)
}
\arguments{
\item{ppm}{concentration of CO2 in ppm}

\item{temp}{temp in celsius}
}
\description{
Convert from ppm to mg/m^3
}
