% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/secant_start.R
\name{secant_starter}
\alias{secant_starter}
\title{secant_starter() Find appropriate starting values for the secant method based on parameter to estimate.}
\usage{
secant_starter(aer = NA, n = NA, Gp = NA, V = NA, C0, C1, delta_t, Cr = 400)
}
\arguments{
\item{aer}{: Air Exchange Rate ( 1 / hour)}

\item{n}{: number of individuals}

\item{Gp}{: Average age-adjusted CO2 generation rate (L / (min·person))}

\item{V}{: Room volume in meters cubed}

\item{C0}{: Initial CO2 concentration (ppm)}

\item{C1}{: Final CO2 concentration (ppm)}

\item{delta_t}{: Time difference (hours)}

\item{Cr}{: Replacement Air CO2 concentration (ppm). Will set to a default value of ambient air (400 ppm)}
}
\value{
A list with the following attributes:
       *start : start value for the secant method
       *stop : stop value for the secant method
       *asymptote : value of x-axis asymptote. Mainly used for debugging. If estimating parameters in the denominator
       of the temp variable in build_up(), Volume or Air Exchange Rate, x-axis value must be selected to be smaller
       than the asymptote. If estimating values in the numerator of temp, number of individuals (n) or Gp, x-axis
       value must be selected to be larger than the asymptote.
}
\description{
secant_starter() Find appropriate starting values for the secant method based on parameter to estimate.
}
