% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate CO2 time series data}
\usage{
simulate_data(
  persondata,
  volume,
  ventilation_rate,
  envCO2 = 400,
  startCO2 = 400,
  freq = 1,
  CO2var = 0.01,
  temp = 25,
  method = "Euler",
  useCpp = TRUE
)
}
\arguments{
\item{persondata}{data.frame with time (in hours since start of measurement), n (number of people), and the following optional columns: age, gender, MET, CO2rate in L/s at 1 atm & 0 Celsius. If you have Gp instead of CO2rate, call Gp_to_CO2rate first}

\item{volume}{volume of room in m^3}

\item{ventilation_rate}{ventilation rate of room in m^3/s}

\item{envCO2}{environmental CO2 concentration in ppm}

\item{startCO2}{starting CO2 concentration in ppm}

\item{freq}{frequency that data should be "measured", in seconds}

\item{CO2var}{variance of noise added to simulated CO2 values at each step}

\item{temp}{temperature of room in Celsius}

\item{method}{method for simulating CO2 concentration. Options are 'Euler' or 'Exponential'}

\item{useCpp}{whether to use Rcpp for simulation. Default is TRUE}
}
\value{
simulated data
}
\description{
Simulate CO2 time series data
}
\examples{
inputdata_example <- data.frame(
  time = c(0, 8, 8, 17, 24),
  n = c(0, 15, 2, 0, 0),
  age = c(0, 8, 50, 0, 0),
  gender = rep(NA, 5),
  MET = c(NA, 3, 1.8, NA, NA),
  CO2rate = rep(NA, 5)
)

exdata <- simulate_data(persondata = inputdata_example, volume=500, ventilation_rate = .1, CO2var = 0.01)
exdata2 <- simulate_data(persondata = inputdata_example, volume=500, ventilation_rate = .5, CO2var = 0.01)
}
