% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transient_mass_balance.R
\name{transient_mass_balance}
\alias{transient_mass_balance}
\title{Estimating ventilation rate, emission rates, and environmental CO2 with the Transient Mass Balance Method}
\usage{
transient_mass_balance(
  freq,
  CO2,
  volume,
  envCO2known = NULL,
  init.Q,
  temp = 25,
  persondata = NULL,
  critpoints = NULL,
  method = "NR",
  max.iter = 1000,
  tol = 1e-08,
  E.init = NULL,
  envCO2.init = NULL,
  envCO2LB = 375,
  envCO2UB = 425,
  ELB = 0,
  EUB = 1000,
  QLB = 0.001,
  QUB = 10,
  verbose = FALSE,
  record.steps = FALSE
)
}
\arguments{
\item{freq}{Frequency of CO2 measurements in seconds}

\item{CO2}{vector of CO2 measurements from sensor in ppm}

\item{volume}{Volume of room in m^3}

\item{envCO2known}{known environmental CO2 concentration in ppm, must be provided for Newton-Raphson (1D optimization)}

\item{init.Q}{initial guess for ventilation rate in m^3/s}

\item{temp}{temperature in degrees Celsius}

\item{persondata}{default NULL, data.frame with time (in hours since start of measurement), n (number of people), and the following optional columns: age, gender, MET, CO2rate in L/s at 1 atm & 0 Celsius. Must be provided for 1D optimization}

\item{critpoints}{default NULL, vector of indices corresponding to the CO2 vector at which to change the estimate of emission rate for multi-dimensional optimization}

\item{method}{'NR' (Newton Raphson) for 1D optimization; 'Newton', 'L-BFGS-B', 'Nelder-Mead' for multidimensional optimization}

\item{max.iter}{Maximum number of iterations for Newton methods}

\item{tol}{Tolerance for Newton methods}

\item{E.init}{initial guess for emission rates in milligrams/s (must be the length of critpoints + 1)}

\item{envCO2.init}{initial guess for environmental CO2 concentration in ppm}

\item{envCO2LB}{for L-BFGS-B, lower bound for environmental CO2 concentration in ppm}

\item{envCO2UB}{for L-BFGS-B, upper bound for environmental CO2 concentration in ppm}

\item{ELB}{for L-BFGS-B, lower bound for emission rates in milligrams/s}

\item{EUB}{for L-BFGS-B, upper bound for emission rates in milligrams/s}

\item{QLB}{for L-BFGS-B, lower bound for ventilation rate in m^3/s}

\item{QUB}{for L-BFGS-B, upper bound for ventilation rate in m^3/s}

\item{record.steps}{}
}
\value{
list with elements: Q (ventilation rate in m^3/s),
ACH (ventilation rate in 1/h),
E (emission rates in milligrams/s) if estimated,
f (value of sum of squares at convergence),
envCO2 (environmental CO2 concentration in ppm) if estimated,
iter (number of iterations),
convergence (convergence code, 0 for success)
}
\description{
Estimating ventilation rate, emission rates, and environmental CO2 with the Transient Mass Balance Method
}
